refactor: refactored module structure. Created a webpack.config and updated package.json. Created folder moduls in folder js and created bundle.js in folder dist
