+++ Food/db.json
@@ -47,6 +47,11 @@
       "name": "sda",
       "phone": "21312",
       "id": 5
+    },
+    {
+      "name": "sad",
+      "phone": "213123",
+      "id": 6
     }
   ]
 }
