+++ Food/index.html
@@ -359,6 +359,6 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script type="module" src="js/app.js"></script>
+    <script type="module" src="dist/js/bundle.js"></script>
   </body>
 </html>
