+++ Food/js/index.js
@@ -1,0 +1,16 @@
+'use strict';
+
+window.addEventListener('DOMContentLoaded', () => {
+  const tabs = require('./modules/tabs'),
+    modal = require('./modules/modal'),
+    timer = require('./modules/timer'),
+    cards = require('./modules/cards'),
+    calc = require('./modules/calc'),
+    slider = require('./modules/slider');
+  tabs();
+  modal();
+  timer();
+  cards();
+  calc();
+  slider();
+});
