+++ Food/webpack.config.js
@@ -1,0 +1,17 @@
+'use strict';
+
+let path = require('path');
+
+module.exports = {
+  mode: 'development',
+  entry: './js/index.js',
+  output: {
+    filename: 'bundle.js',
+    path: __dirname + '/dist/js',
+  },
+  watch: true,
+
+  devtool: 'source-map',
+
+  module: {},
+};
